[fea-rs] Add PendingCompilation (merge support)
This refactors fea-rs to introduce the idea of a PendingCompilation,
which stores all of the state accumulated during the compilation pass of
a single FEA source tree, without finalizing it into OpenType tables.

This is intended to be the basis of support for merging per-master
non-variable FEA into a variable font.
