+++ fea-rs/src/compile/output.rs
@@ -1,11 +1,20 @@
 //! The result of a compilation
 
-use std::collections::HashMap;
+use std::collections::{BTreeMap, HashMap};
 
+use smol_str::SmolStr;
 use write_fonts::{
     BuilderError, FontBuilder,
     tables::{
-        self as wtables, gdef::GlyphClassDef, layout::FeatureParams, maxp::Maxp, stat::AxisValue,
+        self as wtables,
+        gdef::GlyphClassDef,
+        layout::{
+            ConditionSet, FeatureParams, FeatureVariations,
+            builders::CaretValueBuilder as CaretValue,
+        },
+        maxp::Maxp,
+        stat::AxisValue,
+        variations::{common_builder::RemapVarStore, ivs_builder::VariationStoreBuilder},
     },
     types::{GlyphId16, NameId, Tag},
 };
@@ -14,7 +23,39 @@ use super::Opts;
 use super::feature_writer::InsertionPoint;
 use super::tables::DEBG_TAG;
 
-use crate::GlyphMap;
+use crate::{
+    Diagnostic, GlyphMap, GlyphSet, ParseTree,
+    common::MarkClass,
+    compile::{
+        FeatureBuilder, FeatureProvider,
+        features::{AllFeatures, ConditionSetMap},
+        language_system::DefaultLanguageSystems,
+        lookups::{AllLookups, FilterSetId},
+        tables::Tables,
+    },
+};
+
+/// The raw state after compiling one FEA source, before running a feature
+/// provider (if present) and before finalizing.
+///
+/// This exists to support the case of merging two or more different non-variable
+/// FEA sources into a variable output, in the vein of python's varLib.
+pub struct PendingCompilation {
+    pub(super) tree: ParseTree,
+    pub(super) features: AllFeatures,
+    pub(super) lookups: AllLookups,
+    pub(super) tables: Tables,
+    pub(super) default_lang_systems: DefaultLanguageSystems,
+    pub(super) conditionset_defs: ConditionSetMap,
+    pub(super) insert_markers: HashMap<Tag, InsertionPoint>,
+    pub(super) mark_filter_sets: HashMap<GlyphSet, FilterSetId>,
+    pub(super) mark_classes: HashMap<SmolStr, MarkClass>,
+    pub(super) mark_attach_class_id: HashMap<GlyphSet, u16>,
+    pub(super) errors: Vec<Diagnostic>,
+    pub(super) opts: Opts,
+    pub(super) lig_carets_from_feature_writer: BTreeMap<GlyphId16, Vec<CaretValue>>,
+    pub(super) axis_count: u16,
+}
 
 /// The tables generated by this compilation.
 ///
@@ -72,6 +113,202 @@ pub struct Compilation {
     ///
     /// The content is UTF-8 JSON compatible with fontTools' feaLib Debg format.
     pub debg: Option<Vec<u8>>,
+}
+
+impl PendingCompilation {
+    pub(crate) fn run_feature_provider<T: FeatureProvider>(&mut self, provider: &T) {
+        let mut builder = FeatureBuilder::new(
+            &self.default_lang_systems,
+            &mut self.tables,
+            &mut self.mark_filter_sets,
+        );
+        provider.add_features(&mut builder);
+        let mut external_features = builder.finish();
+
+        // we need to register any ConditionSets here, because we want a
+        // stable sort order when we compile
+        if let Some(conditions) = external_features.feature_variations.as_ref() {
+            for conditionset in conditions.conditions.iter().map(|(cs, _)| cs) {
+                self.conditionset_defs.register_use(conditionset);
+            }
+        }
+        external_features.merge_into(&mut self.lookups, &mut self.features, &self.insert_markers);
+        //we need these for build()
+        self.lig_carets_from_feature_writer = external_features.lig_carets;
+    }
+
+    /// Infer/update GDEF table as required.
+    ///
+    /// If a GDEF table is not explicitly defined, we are supposed to create one,
+    /// and even if a GDEF table *is* defined, we are supposed to compute certain
+    /// of its subtables based on other items encountered in the feature file
+    ///
+    /// References:
+    ///
+    /// <http://adobe-type-tools.github.io/afdko/OpenTypeFeatureFileSpecification.html#4f-markclass>
+    /// <http://adobe-type-tools.github.io/afdko/OpenTypeFeatureFileSpecification.html#9b-gdef-table>
+    fn finalize_gdef_table(&mut self) {
+        // if the FEA included a GDEF block, use that, otherwise create an empty table
+        let mut gdef = self.tables.gdef.take().unwrap_or_default();
+        // infer glyph classes, if they were not declared explicitly
+        if gdef.glyph_classes.is_empty() {
+            gdef.glyph_classes_were_inferred = true;
+            self.lookups.infer_glyph_classes(|glyph, class_id| {
+                gdef.glyph_classes.insert(glyph, class_id);
+            });
+            for glyph in self
+                .mark_classes
+                .values()
+                .flat_map(|class| class.members.iter().map(|(cls, _)| cls.iter()))
+                .flatten()
+            {
+                gdef.glyph_classes.insert(glyph, GlyphClassDef::Mark);
+            }
+        }
+
+        if !self.mark_attach_class_id.is_empty() {
+            gdef.mark_attach_class.extend(
+                self.mark_attach_class_id
+                    .iter()
+                    .flat_map(|(cls, id)| cls.iter().map(|gid| (gid, *id))),
+            );
+        }
+
+        if !self.mark_filter_sets.is_empty() {
+            let mut sorted = self
+                .mark_filter_sets
+                .iter()
+                .map(|(cls, id)| (*id, cls.clone()))
+                .collect::<Vec<_>>();
+            sorted.sort_unstable();
+            gdef.mark_glyph_sets = sorted.into_iter().map(|(_, cls)| cls).collect();
+        }
+
+        if gdef.ligature_pos.is_empty() {
+            gdef.ligature_pos = self.lig_carets_from_feature_writer.clone();
+        }
+
+        if !gdef.is_empty() {
+            self.tables.gdef = Some(gdef);
+        }
+    }
+
+    pub(crate) fn build(mut self) -> Result<(Compilation, Vec<Diagnostic>), Vec<Diagnostic>> {
+        self.finalize_gdef_table();
+        self.features
+            .finalize_aalt(&mut self.lookups, &self.default_lang_systems);
+        self.features.dedupe_lookups();
+
+        if self.errors.iter().any(Diagnostic::is_error) {
+            return Err(self.errors.clone());
+        }
+
+        let mut name_builder = self.tables.name.clone();
+        let stat = self
+            .tables
+            .stat
+            .as_ref()
+            .map(|raw| raw.build(&mut name_builder));
+
+        // the var store builder is required so that variable metrics/anchors
+        // in the GPOS table can be collected into an ItemVariationStore
+        let mut ivs = VariationStoreBuilder::new(self.axis_count);
+
+        let (mut gsub, mut gpos) = self.lookups.build(&self.features, &mut ivs, &self.opts);
+        // if ivs hasn't been used, we don't want to create a GDEF table just for it.
+        if !ivs.is_empty() {
+            self.tables
+                .gdef
+                .get_or_insert_with(|| {
+                    // If we're creating a new GdefBuilder here, it means
+                    // finalize_gdef_table() discarded the previous one because
+                    // it was empty. That means no explicit glyph classes were
+                    // declared in the FEA, so mark the classes as inferred.
+                    // https://github.com/googlefonts/fontc/issues/1847
+                    super::tables::GdefBuilder {
+                        glyph_classes_were_inferred: true,
+                        ..Default::default()
+                    }
+                })
+                .var_store = Some(ivs);
+        // but if we _do_ have a gdef table, always add the var store,
+        // since we might still add ligature carets to it
+        } else if let Some(gdef) = self.tables.gdef.as_mut() {
+            gdef.var_store = Some(ivs);
+        }
+
+        let (gdef, key_map) = match self.tables.gdef.as_ref().map(|raw| raw.build()) {
+            Some((gdef, key_map)) => (Some(gdef), key_map),
+            None => (None, None),
+        };
+
+        let feature_params = self.features.build_feature_params(&mut name_builder);
+
+        if let Some(gsub) = gsub.as_mut() {
+            if let Some(variations) = gsub.feature_variations.as_mut() {
+                sort_feature_variations(variations, |condset| {
+                    self.conditionset_defs.sort_order(condset)
+                });
+            }
+            for record in gsub.feature_list.feature_records.iter_mut() {
+                if let Some(params) = feature_params.get(&record.feature_tag) {
+                    record.feature.feature_params = params.clone().into();
+                }
+            }
+        }
+        if let Some(gpos) = gpos.as_mut() {
+            if let Some(key_map) = key_map {
+                // all VariationIndex tables (in value records and anchors)
+                // currently have temporary indices; now that we've built the
+                // ItemVariationStore we need to go and update them all.
+                gpos.remap_variation_indices(&key_map);
+            }
+            if let Some(variations) = gpos.feature_variations.as_mut() {
+                sort_feature_variations(variations, |condset| {
+                    self.conditionset_defs.sort_order(condset)
+                });
+            }
+
+            for record in gpos.feature_list.feature_records.iter_mut() {
+                if let Some(params) = feature_params.get(&record.feature_tag) {
+                    record.feature.feature_params = params.clone().into();
+                }
+            }
+        }
+
+        let gdef_classes = self.tables.gdef.as_ref().and_then(|gdef| {
+            (!gdef.glyph_classes_were_inferred).then(|| gdef.glyph_classes.clone())
+        });
+
+        if self.opts.compile_debg {
+            let (gsub_info, gpos_info) = self.lookups.debug_info();
+            self.tables.debg = Some(super::tables::DebgBuilder::new(
+                gsub_info.to_vec(),
+                gpos_info.to_vec(),
+            ));
+        }
+
+        Ok((
+            Compilation {
+                head: self.tables.head.as_ref().map(|raw| raw.build(None)),
+                hhea: self.tables.hhea.clone(),
+                vhea: self.tables.vhea.clone(),
+                os2: self.tables.os2.as_ref().map(|raw| raw.build()),
+                os2_builder: self.tables.os2.clone(),
+                gdef,
+                base: self.tables.base.as_ref().map(|raw| raw.build()),
+                name: name_builder.build(),
+                stat,
+                gsub,
+                gpos,
+                opts: self.opts.clone(),
+                gdef_classes,
+                insert_markers: self.insert_markers.clone(),
+                debg: self.tables.debg.as_ref().map(|d| d.build(&self.tree)),
+            },
+            self.errors.clone(),
+        ))
+    }
 }
 
 impl Compilation {
@@ -216,4 +453,16 @@ impl Compilation {
 
         Ok(builder.build())
     }
+}
+
+fn sort_feature_variations(
+    variations: &mut FeatureVariations,
+    order_fn: impl Fn(&ConditionSet) -> usize,
+) {
+    variations
+        .feature_variation_records
+        .sort_by_key(|record| match record.condition_set.as_ref() {
+            Some(condition) => order_fn(condition),
+            None => order_fn(&Default::default()),
+        })
 }
